function checked done error handling done TODO app done

+++ Todo_app/TODO.py
@@ -7,18 +7,26 @@ class Controller():
         model = Model()
         if len(sys.argv) == 1:
             display.show_menu()
-        elif sys.argv[1] == "-l":
-            model.print_file()
-        elif sys.argv[1] == "-a":
-            if sys.argv[2:] == []:
-                print("Unable to add: no task provided")
+        else:
+            if sys.argv[1] == "-l" and sys.argv[2:] == []:
+                model.print_file()
+            elif sys.argv[1] == "-a":
+                if sys.argv[2:] == []:
+                    print("Unable to add: no task provided")
+                else:
+                    model.add_new_task()
+            elif sys.argv[1] == "-r":
+                if sys.argv[2:] == []:
+                    print("Unable to remove: no index provided")
+                else:
+                    model.remove_task()
+            elif sys.argv[1] == "-c":
+                if sys.argv[2:] == []:
+                    print("Unable to check: no index provided")
+                else:
+                    model.check_task()
             else:
-                model.add_new_task()
-        elif sys.argv[1] == "-r":
-            if sys.argv[2:] == []:
-                print("Unable to remove: no index provided")
-            else:
-                model.remove_task()
+                print("Unsupported argument")
 
 class Model():
     def __init__(self):
@@ -40,7 +48,7 @@ class Model():
         text_append = " ".join(sys.argv[2:])
         if os.path.exists(self.file_name):
             with open(self.file_name, 'a') as file:
-                file.write(text_append + "\n")
+                file.write("[ ] " + text_append + "\n")
 
     def remove_task(self):
         try:
@@ -54,6 +62,22 @@ class Model():
                     print("Unable to remove: index is out of bound")
                 else:
                     del lines[delete_line_index - 1]
+                    open_file.seek(0)
+                    open_file.truncate()
+                    open_file.writelines(lines)
+
+    def check_task(self):
+        try:
+            delete_line_index = int(sys.argv[2])
+        except ValueError:
+            print("Unable to check: index is not a number")
+        else:
+            with open(self.file_name, "r+") as open_file:
+                lines = open_file.readlines()
+                if len(lines) < delete_line_index:
+                    print("Unable to check: index is out of bound")
+                else:
+                    lines[delete_line_index - 1] = "[x]" + lines[delete_line_index - 1][3:]
                     open_file.seek(0)
                     open_file.truncate()
                     open_file.writelines(lines)
